requirements.txt
django
python-dateutil
pyyaml
requests
dkimpy
mssql-django
mysqlclient
shortuuid
